New db layout once more; thumbnails are files
Now, thumbnails are also listed as files on the db, and their filename
on disk has more info. Queries are now more complicated, but this is for
good (I think)

When developing this, I put this on ts:

export PGDATABASE=dev_testing_risos

and compiled with PGDATABASE=dev_testing_risos omake

(I did an alias for that, git t)

That way I could test without wiping the previous dev repository

I'm removing this from the commit, but seems nice. But putting this to
tagsystem user seems better anyway.

Will try to merge this feature branch on master, then merge master on
stable, then rename master to devel. Or maybe merge stable on master,
drop stable, rename master to stable.. etc.